Move modvendor tool to fyne sub command
